Increase frame buffer to prevent frame loss with new ipecamera with fast DMA
The hardware is able to store about 200 frames. If hardware buffer is full,
the data will be sent with the speed of DMA engine. This can be faster 
when expected rate of 300 frames per second. It may be not possible to 
decode frames at such speed and unless all frames are fit into the buffer,
the frame lost will occur.